streamingccl: delete CREATE REPLICATION STREAM implementation
CREATE REPLICATION STREAM is the old style replication stream
implementation which uses CDC and is deprecated.

Release justification: Cat 1.
Release note: None
