+++ pkg/ccl/streamingccl/streamproducer/BUILD.bazel
@@ -6,14 +6,11 @@ go_library(
         "event_stream.go",
         "producer_job.go",
         "replication_manager.go",
-        "replication_stream_planning.go",
         "stream_lifetime.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/streamproducer",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/ccl/changefeedccl/changefeedbase",
-        "//pkg/ccl/changefeedccl/changefeeddist",
         "//pkg/ccl/streamingccl",
         "//pkg/ccl/streamingccl/streampb",
         "//pkg/ccl/utilccl",
@@ -27,10 +24,8 @@ go_library(
         "//pkg/kv/kvserver/protectedts/ptpb",
         "//pkg/roachpb",
         "//pkg/security",
-        "//pkg/server/telemetry",
         "//pkg/settings/cluster",
         "//pkg/sql",
-        "//pkg/sql/catalog/colinfo",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/sem/tree",
@@ -86,7 +81,6 @@ go_test(
         "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondatapb",
         "//pkg/streaming",
-        "//pkg/testutils",
         "//pkg/testutils/serverutils",
         "//pkg/testutils/sqlutils",
         "//pkg/testutils/testcluster",
