+++ pkg/gen/docs.bzl
@@ -88,7 +88,6 @@ DOCS_SRCS = [
   "//docs/generated/sql/bnf:create_extension_stmt.bnf",
   "//docs/generated/sql/bnf:create_index_stmt.bnf",
   "//docs/generated/sql/bnf:create_inverted_index_stmt.bnf",
-  "//docs/generated/sql/bnf:create_replication_stream_stmt.bnf",
   "//docs/generated/sql/bnf:create_role_stmt.bnf",
   "//docs/generated/sql/bnf:create_schedule_for_backup_stmt.bnf",
   "//docs/generated/sql/bnf:create_schema_stmt.bnf",
