+++ pkg/sql/sem/tree/BUILD.bazel
@@ -82,7 +82,6 @@ go_library(
         "regexp_cache.go",
         "region.go",
         "rename.go",
-        "replication_stream.go",
         "returning.go",
         "revoke.go",
         "role_spec.go",
